E2E test support definition of test files for each version
