+++ test/e2e/config/docker.yaml
@@ -65,13 +65,13 @@ providers:
     replacements:
     - old: --metrics-bind-addr=127.0.0.1:8080
       new: --metrics-bind-addr=:8080
-  files:
-  # Add cluster templates
-  - sourcePath: "../data/infrastructure-docker/v1alpha4/cluster-template.yaml"
-  - sourcePath: "../data/infrastructure-docker/v1alpha4/cluster-template-mhc.yaml"
-  - sourcePath: "../data/infrastructure-docker/v1alpha4/cluster-template-kcp-adoption.yaml"
-  - sourcePath: "../data/infrastructure-docker/v1alpha4/cluster-template-machine-pool.yaml"
-  - sourcePath: "../data/infrastructure-docker/v1alpha4/cluster-template-node-drain.yaml"
+    files:
+    # Add cluster templates
+    - sourcePath: "../data/infrastructure-docker/v1alpha4/cluster-template.yaml"
+    - sourcePath: "../data/infrastructure-docker/v1alpha4/cluster-template-mhc.yaml"
+    - sourcePath: "../data/infrastructure-docker/v1alpha4/cluster-template-kcp-adoption.yaml"
+    - sourcePath: "../data/infrastructure-docker/v1alpha4/cluster-template-machine-pool.yaml"
+    - sourcePath: "../data/infrastructure-docker/v1alpha4/cluster-template-node-drain.yaml"
 
 variables:
   KUBERNETES_VERSION: "v1.19.1"
