+++ test/framework/clusterctl/repository.go
@@ -21,15 +21,18 @@ import (
 	"context"
 	"fmt"
 	"io/ioutil"
+	"net/http"
 	"os"
 	"path/filepath"
+	"regexp"
 	"strings"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"github.com/pkg/errors"
+	"sigs.k8s.io/cluster-api/test/framework/exec"
 
 	clusterctlv1 "sigs.k8s.io/cluster-api/cmd/clusterctl/api/v1alpha3"
-	"sigs.k8s.io/cluster-api/test/framework"
 )
 
 // Provides helpers for managing a clusterctl local repository to be used for running e2e tests in isolation.
@@ -71,12 +74,10 @@ func CreateRepository(ctx context.Context, input CreateRepositoryInput) string {
 
 	providers := []providerConfig{}
 	for _, provider := range input.E2EConfig.Providers {
-		providerURL := ""
+		providerLabel := clusterctlv1.ManifestLabel(provider.Name, clusterctlv1.ProviderType(provider.Type))
+		providerURL := filepath.Join(input.RepositoryFolder, providerLabel, "latest", "components.yaml")
 		for _, version := range provider.Versions {
-			providerLabel := clusterctlv1.ManifestLabel(provider.Name, clusterctlv1.ProviderType(provider.Type))
-
-			generator := framework.ComponentGeneratorForComponentSource(version)
-			manifest, err := generator.Manifests(ctx)
+			manifest, err := YAMLForComponentSource(ctx, version)
 			Expect(err).ToNot(HaveOccurred(), "Failed to generate the manifest for %q / %q", providerLabel, version.Name)
 
 			sourcePath := filepath.Join(input.RepositoryFolder, providerLabel, version.Name)
@@ -85,8 +86,20 @@ func CreateRepository(ctx context.Context, input CreateRepositoryInput) string {
 			filePath := filepath.Join(sourcePath, "components.yaml")
 			Expect(ioutil.WriteFile(filePath, manifest, 0600)).To(Succeed(), "Failed to write manifest in the clusterctl local repository for %q / %q", providerLabel, version.Name)
 
-			if providerURL == "" {
-				providerURL = filePath
+			destinationPath := filepath.Join(input.RepositoryFolder, providerLabel, version.Name, "components.yaml")
+			allFiles := append(provider.Files, version.Files...)
+			for _, file := range allFiles {
+				data, err := ioutil.ReadFile(file.SourcePath)
+				Expect(err).ToNot(HaveOccurred(), "Failed to read file %q / %q", provider.Name, file.SourcePath)
+
+				// Applies FileTransformations if defined
+				for _, t := range input.FileTransformations {
+					data, err = t(data)
+					Expect(err).ToNot(HaveOccurred(), "Failed to apply transformation func template %q", file)
+				}
+
+				destinationFile := filepath.Join(filepath.Dir(destinationPath), file.TargetName)
+				Expect(ioutil.WriteFile(destinationFile, data, 0600)).To(Succeed(), "Failed to write clusterctl local repository file %q / %q", provider.Name, file.TargetName)
 			}
 		}
 		providers = append(providers, providerConfig{
@@ -94,20 +107,6 @@ func CreateRepository(ctx context.Context, input CreateRepositoryInput) string {
 			URL:  providerURL,
 			Type: provider.Type,
 		})
-
-		for _, file := range provider.Files {
-			data, err := ioutil.ReadFile(file.SourcePath)
-			Expect(err).ToNot(HaveOccurred(), "Failed to read file %q / %q", provider.Name, file.SourcePath)
-
-			// Applies FileTransformations if defined
-			for _, t := range input.FileTransformations {
-				data, err = t(data)
-				Expect(err).ToNot(HaveOccurred(), "Failed to apply transformation func template %q", file)
-			}
-
-			destinationFile := filepath.Join(filepath.Dir(providerURL), file.TargetName)
-			Expect(ioutil.WriteFile(destinationFile, data, 0600)).To(Succeed(), "Failed to write clusterctl local repository file %q / %q", provider.Name, file.TargetName)
-		}
 	}
 
 	// set this path to an empty file under the repository path, so test can run in isolation without user's overrides kicking in
@@ -128,4 +127,44 @@ func CreateRepository(ctx context.Context, input CreateRepositoryInput) string {
 	clusterctlConfigFile.write()
 
 	return clusterctlConfigFile.Path
+}
+
+// YAMLForComponentSource returns the YAML for the provided component source.
+func YAMLForComponentSource(ctx context.Context, source ProviderVersionSource) ([]byte, error) {
+	var data []byte
+
+	switch source.Type {
+	case URLSource:
+		resp, err := http.Get(source.Value)
+		if err != nil {
+			return nil, err
+		}
+		defer resp.Body.Close()
+		buf, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			return nil, err
+		}
+		data = buf
+	case KustomizeSource:
+		kustomize := exec.NewCommand(
+			exec.WithCommand("kustomize"),
+			exec.WithArgs("build", source.Value))
+		stdout, stderr, err := kustomize.Run(ctx)
+		if err != nil {
+			return nil, errors.Wrapf(err, "failed to execute kustomize: %s", stderr)
+		}
+		data = stdout
+	default:
+		return nil, errors.Errorf("invalid type: %q", source.Type)
+	}
+
+	for _, replacement := range source.Replacements {
+		rx, err := regexp.Compile(replacement.Old)
+		if err != nil {
+			return nil, err
+		}
+		data = rx.ReplaceAll(data, []byte(replacement.New))
+	}
+
+	return data, nil
 }
